Use embed in internal profile
