+++ internal/profile/_static/package_registry.yml
@@ -1,0 +1,5 @@
+package_paths:
+  - /packages/development
+  - /packages/production
+  - /packages/staging
+  - /packages/snapshot
